--- ee/app/engagement-dashboard/client/components/MessagesTab/index.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Divider } from '@rocket.chat/fuselage';
-
-import MessagesSentSection from './MessagesSentSection';
-import MessagesPerChannelSection from './MessagesPerChannelSection';
-
-const MessagesTab = () => <>
-	<MessagesSentSection />
-	<Divider />
-	<MessagesPerChannelSection />
-</>;
-
-export default MessagesTab;
