--- ee/app/engagement-dashboard/client/components/data/colors.js
@@ -1,2 +1,0 @@
-export const monochromaticColors = ['#E8F2FF', '#D1EBFE', '#A4D3FE', '#76B7FC', '#549DF9', '#1D74F5', '#10529E'];
-export const polychromaticColors = ['#FFD031', '#2DE0A5', '#1D74F5'];
