--- ee/app/engagement-dashboard/client/index.js
@@ -1,1 +1,0 @@
-import './routes';
