+++ ee/client/index.ts
@@ -1,7 +1,7 @@
 import '../app/auditing/client';
 import '../app/authorization/client';
 import '../app/canned-responses/client';
-import '../app/engagement-dashboard/client';
 import '../app/license/client';
 import '../app/livechat-enterprise/client';
 import './omnichannel';
+import './startup';
