+++ ee/client/startup/index.ts
@@ -1,0 +1,1 @@
+import './engagementDashboard';
