+++ ee/client/views/admin/engagementDashboard/ChannelsTab/ChannelsTab.stories.js
@@ -6,9 +6,7 @@ import ChannelsTab from '.';
 export default {
 	title: 'admin/enterprise/engagement/ChannelsTab',
 	component: ChannelsTab,
-	decorators: [
-		(fn) => <Margins children={fn()} all='x24' />,
-	],
+	decorators: [(fn) => <Margins children={fn()} all='x24' />],
 };
 
 export const _default = () => <ChannelsTab />;
