+++ ee/client/views/admin/engagementDashboard/MessagesTab/MessagesTab.stories.js
@@ -6,9 +6,7 @@ import MessagesTab from '.';
 export default {
 	title: 'admin/enterprise/engagement/MessagesTab',
 	component: MessagesTab,
-	decorators: [
-		(fn) => <Margins children={fn()} all='x24' />,
-	],
+	decorators: [(fn) => <Margins children={fn()} all='x24' />],
 };
 
 export const _default = () => <MessagesTab />;
