+++ ee/client/views/admin/engagementDashboard/UsersTab/UsersTab.Stories.js
@@ -1,14 +1,12 @@
 import { Margins } from '@rocket.chat/fuselage';
 import React from 'react';
 
-import { UsersTab } from '.';
+import UsersTab from '.';
 
 export default {
 	title: 'admin/enterprise/engagement/UsersTab',
 	component: UsersTab,
-	decorators: [
-		(fn) => <Margins children={fn()} all='x24' />,
-	],
+	decorators: [(fn) => <Margins children={fn()} all='x24' />],
 };
 
 export const _default = () => <UsersTab />;
