+++ ee/client/views/admin/engagementDashboard/data/Histogram.stories.js
@@ -6,11 +6,15 @@ import { Histogram } from './Histogram';
 export default {
 	title: 'admin/enterprise/engagement/data/Histogram',
 	component: Histogram,
-	decorators: [(fn) => <Margins all='x16'>
-		<Flex.Container>
-			<Box children={fn()} style={{ height: 240 }} />
-		</Flex.Container>
-	</Margins>],
+	decorators: [
+		(fn) => (
+			<Margins all='x16'>
+				<Flex.Container>
+					<Box children={fn()} style={{ height: 240 }} />
+				</Flex.Container>
+			</Margins>
+		),
+	],
 };
 
 export const _default = () => <Histogram />;
